requirements.txt
Pillow
PyMuPDF
brotli
tqdm
tkinterdnd2
